Move ImageFormat to screenshotter
