+++ common/screenshotter.go
@@ -1,7 +1,9 @@
 package common
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
@@ -15,6 +17,49 @@ import (
 	cdppage "github.com/chromedp/cdproto/page"
 	"github.com/dop251/goja"
 )
+
+// ImageFormat represents an image file format.
+type ImageFormat string
+
+// Valid image format options.
+const (
+	ImageFormatJPEG ImageFormat = "jpeg"
+	ImageFormatPNG  ImageFormat = "png"
+)
+
+func (f ImageFormat) String() string {
+	return imageFormatToString[f]
+}
+
+var imageFormatToString = map[ImageFormat]string{ //nolint:gochecknoglobals
+	ImageFormatJPEG: "jpeg",
+	ImageFormatPNG:  "png",
+}
+
+var imageFormatToID = map[string]ImageFormat{ //nolint:gochecknoglobals
+	"jpeg": ImageFormatJPEG,
+	"png":  ImageFormatPNG,
+}
+
+// MarshalJSON marshals the enum as a quoted JSON string.
+func (f ImageFormat) MarshalJSON() ([]byte, error) {
+	buffer := bytes.NewBufferString(`"`)
+	buffer.WriteString(imageFormatToString[f])
+	buffer.WriteString(`"`)
+	return buffer.Bytes(), nil
+}
+
+// UnmarshalJSON unmarshals a quoted JSON string to the enum value.
+func (f *ImageFormat) UnmarshalJSON(b []byte) error {
+	var j string
+	err := json.Unmarshal(b, &j)
+	if err != nil {
+		return fmt.Errorf("unmarshalling image format: %w", err)
+	}
+	// Note that if the string cannot be found then it will be set to the zero value.
+	*f = imageFormatToID[j]
+	return nil
+}
 
 type screenshotter struct {
 	ctx context.Context
